requirements.txt
asgiref==3.4.1
certifi==2021.5.30
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
drf-nested-routers==0.93.3
drf-yasg==1.20.0
drf-yasg2==1.19.4
environs==9.3.2
idna==2.10
inflection==0.5.1
install==1.3.4
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.12.1
packaging==20.9
psycopg2==2.9.1
PyJWT==2.1.0
pyparsing==2.4.7
python-dotenv==0.18.0
pytz==2021.1
requests==2.25.1
routers==0.10.1
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.4
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
